feat: 게임 실행 컨트롤러 LottoController 추가
